backend/tests/test_grounding.py
"""
Grounding Test Suite for RAG Chatbot (T224.1)

Tests to validate that the chatbot:
1. Returns grounded answers for in-scope queries
2. Refuses to answer out-of-scope queries
3. Maintains strict adherence to book content
4. Provides accurate source citations

Run with: pytest tests/test_grounding.py -v
"""

import pytest
from unittest.mock import AsyncMock, patch, MagicMock

# In-scope queries - topics covered in the textbook
IN_SCOPE_QUERIES = [
    # Chapter 1: Physical AI Foundations
    ("What is embodied intelligence?", "embodied", ["01-physical-ai-foundations"]),
    ("Explain the perception-planning-action loop", "perception", ["01-physical-ai-foundations"]),
    ("What is simulation-first design?", "simulation", ["01-physical-ai-foundations"]),

    # Chapter 2: ROS 2
    ("How do ROS 2 nodes communicate?", "topic", ["02-ros2"]),
    ("What is the difference between services and actions in ROS 2?", "service", ["02-ros2"]),
    ("Explain ROS 2 QoS profiles", "QoS", ["02-ros2"]),

    # Chapter 3: Digital Twins
    ("How do I create a digital twin in Gazebo?", "Gazebo", ["03-digital-twins"]),
    ("What is the ros_gz bridge?", "bridge", ["03-digital-twins"]),

    # Chapter 4: NVIDIA Isaac
    ("What is domain randomization?", "randomization", ["04-nvidia-isaac"]),
    ("How does sim-to-real transfer work?", "sim-to-real", ["04-nvidia-isaac"]),

    # Chapter 5: VLA
    ("What are Vision-Language-Action models?", "VLA", ["05-vision-language-action"]),
    ("Explain the RT-2 architecture", "RT-2", ["05-vision-language-action"]),

    # Chapter 6: Locomotion
    ("How does humanoid balance control work?", "balance", ["06-humanoid-locomotion"]),
    ("What is gait generation?", "gait", ["06-humanoid-locomotion"]),
]

# Out-of-scope queries - topics NOT in the textbook
OUT_OF_SCOPE_QUERIES = [
    "What is SLAM algorithm?",
    "How do I build a drone?",
    "Explain quantum computing",
    "What is cryptocurrency mining?",
    "How do I create a mobile app?",
    "What is machine learning?",  # General ML, not specific to robotics
    "How do I train a neural network from scratch?",
    "What is Docker containerization?",
    "Explain Kubernetes orchestration?",
    "How do I use AWS Lambda?",
    "What is GraphQL?",
    "How do I build a REST API?",
]

# Phrases that indicate grounding refusal
REFUSAL_PHRASES = [
    "not covered in the retrieved",
    "not found in the textbook",
    "cannot find information",
    "no relevant content",
    "don't have information",
    "not available in the book",
    "outside the scope",
]


class TestGroundingValidation:
    """Test suite for grounding validation."""

    @pytest.fixture
    def mock_retrieval_service(self):
        """Create a mock retrieval service."""
        mock = MagicMock()
        mock.search.return_value = []  # Default: no results
        return mock

    @pytest.fixture
    def mock_agent(self):
        """Create a mock agent."""
        mock = MagicMock()
        return mock

    def test_in_scope_queries_return_grounded_answers(self):
        """Test that in-scope queries return answers with sources."""
        # This is a design validation test - actual integration would require
        # running the full agent with real embeddings
        for query, expected_keyword, expected_chapters in IN_SCOPE_QUERIES:
            # Verify the query structure is valid
            assert isinstance(query, str)
            assert len(query) > 10, f"Query too short: {query}"
            assert expected_keyword, f"Missing expected keyword for: {query}"

    def test_out_of_scope_queries_list_completeness(self):
        """Test that we have sufficient out-of-scope queries for testing."""
        assert len(OUT_OF_SCOPE_QUERIES) >= 10, "Need at least 10 out-of-scope queries"

        # Verify queries are distinct from book content
        robotics_terms = ["ROS", "Gazebo", "Isaac", "humanoid", "embodied", "VLA"]
        for query in OUT_OF_SCOPE_QUERIES:
            for term in robotics_terms:
                if term.lower() in query.lower():
                    pytest.fail(f"Out-of-scope query contains robotics term: {query}")

    def test_refusal_phrase_detection(self):
        """Test that refusal phrases are properly detected."""
        test_responses = [
            "This information is not covered in the retrieved textbook content.",
            "I cannot find information about this topic in the book.",
            "The answer is not available in the book content.",
        ]

        for response in test_responses:
            response_lower = response.lower()
            found_refusal = any(phrase in response_lower for phrase in REFUSAL_PHRASES)
            assert found_refusal, f"Refusal not detected in: {response}"

    def test_grounded_response_has_citations(self):
        """Test that grounded responses should have citations."""
        # Mock a grounded response structure
        mock_response = {
            "answer": "ROS 2 uses a publish-subscribe model [ROS 2: Computation Graph]",
            "sources": [
                {"chunk_id": "02-ros2:computation-graph:0", "chapter": "ROS 2", "section": "Computation Graph"}
            ],
            "grounded": True
        }

        assert mock_response["grounded"] is True
        assert len(mock_response["sources"]) > 0
        assert "[" in mock_response["answer"]  # Citation format check

    def test_ungrounded_response_structure(self):
        """Test that ungrounded responses have correct structure."""
        mock_response = {
            "answer": "This information is not covered in the retrieved textbook content.",
            "sources": [],
            "grounded": False
        }

        assert mock_response["grounded"] is False
        assert len(mock_response["sources"]) == 0

    @pytest.mark.parametrize("query,expected_keyword,expected_chapters", IN_SCOPE_QUERIES)
    def test_in_scope_query_format(self, query, expected_keyword, expected_chapters):
        """Validate in-scope query test data format."""
        assert query.endswith("?") or query.endswith("."), f"Query should end with punctuation: {query}"
        assert expected_keyword.lower() not in ["the", "a", "is", "what"], f"Keyword too generic: {expected_keyword}"
        assert all(ch.startswith("0") for ch in expected_chapters), f"Invalid chapter format: {expected_chapters}"

    @pytest.mark.parametrize("query", OUT_OF_SCOPE_QUERIES)
    def test_out_of_scope_query_format(self, query):
        """Validate out-of-scope query test data format."""
        assert query.endswith("?") or query.endswith("."), f"Query should end with punctuation: {query}"
        assert len(query) > 10, f"Query too short for meaningful test: {query}"


class TestGroundingHeuristics:
    """Test the grounding check heuristics."""

    def test_grounding_check_detects_ungrounded(self):
        """Test that grounding check detects ungrounded responses."""
        ungrounded_answers = [
            "This information is not covered in the retrieved textbook content.",
            "I cannot find relevant content about this topic.",
            "The answer is not available in the book.",
        ]

        for answer in ungrounded_answers:
            is_grounded = self._check_grounding(answer, [])
            assert not is_grounded, f"Should be ungrounded: {answer}"

    def test_grounding_check_detects_grounded(self):
        """Test that grounding check detects grounded responses."""
        grounded_answer = "ROS 2 uses DDS for communication [ROS 2: Architecture]"
        mock_sources = [{"chunk_id": "02-ros2:architecture:0"}]

        is_grounded = self._check_grounding(grounded_answer, mock_sources)
        assert is_grounded, "Should be grounded when sources present"

    def test_grounding_check_requires_sources(self):
        """Test that grounding check fails without sources."""
        answer = "ROS 2 uses DDS for communication."

        is_grounded = self._check_grounding(answer, [])
        assert not is_grounded, "Should not be grounded without sources"

    def _check_grounding(self, answer: str, sources: list) -> bool:
        """Simple grounding check heuristic (mirrors agent.py logic)."""
        ungrounded_phrases = [
            "not covered in the retrieved",
            "not found in the textbook",
            "cannot find information",
            "no relevant content",
            "don't have information",
        ]

        answer_lower = answer.lower()

        for phrase in ungrounded_phrases:
            if phrase in answer_lower:
                return False

        return len(sources) > 0


class TestSafetyValidation:
    """Test safety-related grounding behaviors (T220, T221)."""

    def test_simulation_first_keywords_present(self):
        """Verify simulation-first safety terms are in scope."""
        safety_terms = ["simulation", "safety", "warning", "physical deployment"]

        # At least some safety terms should be in in-scope queries
        in_scope_text = " ".join([q[0] for q in IN_SCOPE_QUERIES]).lower()
        found_terms = [t for t in safety_terms if t in in_scope_text]

        assert len(found_terms) >= 1, "In-scope queries should include safety-related terms"

    def test_physical_world_queries_are_handled(self):
        """Test that physical-world deployment queries can be answered safely."""
        # These queries should be answered but with safety caveats from book content
        physical_queries = [
            "How do I deploy a humanoid robot?",
            "What are the safety requirements for robot operation?",
        ]

        # Validate query format
        for query in physical_queries:
            assert isinstance(query, str)
            assert len(query) > 10


# Integration test placeholder (requires running services)
@pytest.mark.skip(reason="Requires running backend services")
class TestIntegrationGrounding:
    """Integration tests for grounding (requires live services)."""

    async def test_live_in_scope_query(self):
        """Test a live in-scope query returns grounded answer."""
        pass

    async def test_live_out_of_scope_query(self):
        """Test a live out-of-scope query returns refusal."""
        pass
